QC-15 Add on change events on booking form
