+++ react/qa-cinema/src/components/bookings/bookings.css
@@ -1,0 +1,3 @@
+.ticket-alert {
+    width: fit-content;
+}
